Use mock(1) to build RPM packages.
Mock(1) builds RPM packages in a chroot environment, containing only
a base system and the build dependencies specified in the RPM spec
file.  This ensures that the spec file is correct, and not missing
any dependencies, and thus avoids "contaminating" the packages with
whatever happened to be installed on the build host.  It is a good
idea to always use mock when building packages, so we here change
the "rpms" target to use mock instead of calling rpmbuild directly.

To build for a different operating system or architecture, specify
the 'target' variable on the command line to make, e.g:

    $ make rpms target=epel-8-x86_64

This passes the $(target) variable to mock's -r option, to select
the mock(1) chroot configuration.

The resulting RPMs, and the mock log files, will be stored in the
"packages/<rootname>" directory, where "<rootname>" is the root
name in the selected mock config (generally the same as the mock
chroot config file name).

(In a default installation, mock will ask for the root password when
run, and we will actually run it twice, so you get asked twice.  Get
added to the "mock" group, or configure PAM to permit you otherwise,
to avoid this.)