hostgrep: dynamically add characters allowed in hostnames.
Until now, hostnames were allowed to contain exactly upper and lower
case A-Z and the digits 0-9. This meant that you could not grep for
names like "rack02-pos13". Change to dynamically add additional
characters from the hostlist provided on the command line, so that it
will work to grep for hostnames containing those characters.

Based on feedback from Rowan Worth <rowanw@dug.com>